Quoll.Parser

TOML input parsing. The Parser submodule is not reexported, so all of its symbols are listed here.

Quoll.Parser.find_leafdirs — Method

find_leafdirs(rootdir::T) where T

Find leaf directories (directories that do not contain directories themselves) going down from rootdir directory. Absolute paths are returned.

Quoll.Parser.normalize_comparison — Method

normalize_comparison(s)

Normalize a string s by making the string lowercase and replacing certain characters (underscores, hyphens and whitespace are removed by default as defined in replace_pairs) for easier comparison with reference values.